requirements.txt
SQLAlchemy==2.0.40
dotenv==0.9.9
fastapi==0.115.8
httpx==0.28.1
psycopg2-binary==2.9.10
pydantic==2.10.6
python-dotenv==1.0.1
